requirements.txt
# Core
numpy>=1.23
pandas>=1.5
matplotlib>=3.7
opencv-python>=4.7
Pillow>=9.5
PyYAML>=6.0
requests>=2.28
scipy>=1.10
tqdm>=4.65

# Deep Learning
torch>=2.0
torchvision>=0.15
ultralytics>=8.0.200

# Utils
psutil
gitpython

# App
flask
flask-cors
gdown

# Optional (safe)
seaborn>=0.12
from_root


-e .